+++ com.unity.netcode.gameobjects/Runtime/Components/AnticipatedNetworkTransform.cs
@@ -1,4 +1,5 @@
 using Unity.Mathematics;
+using Unity.Netcode.GameObjects.Timing;
 using Unity.Netcode.Runtime;
 using UnityEngine;
