+++ com.unity.netcode.gameobjects/Runtime/Components/NetworkTransform.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Text;
 using Unity.Mathematics;
+using Unity.Netcode.GameObjects.Timing;
 using Unity.Netcode.Runtime;
 using UnityEngine;
