+++ com.unity.netcode.gameobjects/Runtime/Core/ComponentFactory.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Unity.Netcode.GameObjects.Timing;
 
 namespace Unity.Netcode
 {
