+++ com.unity.netcode.gameobjects/Runtime/Messaging/Messages/TimeSyncMessage.cs
@@ -1,3 +1,5 @@
+using Unity.Netcode.GameObjects.Timing;
+
 namespace Unity.Netcode
 {
     internal struct TimeSyncMessage : INetworkMessage, INetworkSerializeByMemcpy
