+++ com.unity.netcode.gameobjects/Runtime/Timing/NetworkTimeSystem.cs
@@ -1,8 +1,10 @@
 using System;
+using Unity.Netcode;
 using Unity.Profiling;
 using UnityEngine;
+using UnityEngine.Scripting.APIUpdating;
 
-namespace Unity.Netcode
+namespace Unity.Netcode.GameObjects.Timing
 {
     /// <summary>
     /// <see cref="NetworkTimeSystem"/> is a standalone system which can be used to run a network time simulation.
@@ -11,6 +13,7 @@ namespace Unity.Netcode
     /// effort at predicting what the server tick will be when a given network action is processed on the server.
     /// </summary>
     [Serializable]
+    [MovedFrom(true, "Unity.Netcode", null, null)]
     public class NetworkTimeSystem
     {
         /// <remarks>
@@ -85,12 +88,12 @@ namespace Unity.Netcode
         /// <remarks>
         /// For a distributed authority network topology, this latency is between the client and the
         /// distributed authority service instance.<br />
-        /// Note: <see cref="Components.NetworkTransform"/> uses this value plus an additional global
-        /// offset <see cref="Components.NetworkTransform.InterpolationBufferTickOffset"/> when interpolation
+        /// Note: <see cref="Unity.Netcode.Components.NetworkTransform"/> uses this value plus an additional global
+        /// offset <see cref="Unity.Netcode.Components.NetworkTransform.InterpolationBufferTickOffset"/> when interpolation
         /// is enabled. <br />
-        /// To see the current <see cref="Components.NetworkTransform"/> tick latency: <br />
-        /// - <see cref="Components.NetworkTransform.GetTickLatency"/> <br />
-        /// - <see cref="Components.NetworkTransform.GetTickLatencyInSeconds"/> <br />
+        /// To see the current <see cref="Unity.Netcode.Components.NetworkTransform"/> tick latency: <br />
+        /// - <see cref="Unity.Netcode.Components.NetworkTransform.GetTickLatency"/> <br />
+        /// - <see cref="Unity.Netcode.Components.NetworkTransform.GetTickLatencyInSeconds"/> <br />
         /// </remarks>
         public int TickLatency = 1;
