+++ com.unity.netcode.gameobjects/Tests/Editor/Timing/ClientNetworkTimeSystemTests.cs
@@ -1,5 +1,6 @@
 using System;
 using NUnit.Framework;
+using Unity.Netcode.GameObjects.Timing;
 using UnityEngine;
 
 namespace Unity.Netcode.GameObjects.EditorTests
