+++ com.unity.netcode.gameobjects/Tests/Editor/Timing/NetworkTimeTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using NUnit.Framework;
+using Unity.Netcode.GameObjects.Timing;
 using UnityEngine;
 using Random = System.Random;
