+++ com.unity.netcode.gameobjects/Tests/Editor/Timing/TimingTestHelper.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Unity.Netcode.GameObjects.Timing;
 using UnityEngine;
 using Random = System.Random;
