+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkManagerEventsTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using NUnit.Framework;
+using Unity.Netcode.GameObjects.Timing;
 using Unity.Netcode.TestHelpers.Runtime;
 using UnityEngine;
 using UnityEngine.TestTools;
