+++ com.unity.netcode.gameobjects/Tests/Runtime/TestHelpers/MockTimeProvider.cs
@@ -1,3 +1,5 @@
+using Unity.Netcode.GameObjects.Timing;
+
 namespace Unity.Netcode.TestHelpers.Runtime
 {
     /// <summary>
